Move some http2 specific tools to a public location
