+++ http/src/main/scala/org/http4s/blaze/http/http20/BasicHttpStage.scala
@@ -10,7 +10,6 @@ import org.http4s.blaze.pipeline.TailStage
 import org.http4s.blaze.util.BufferTools
 import Http2Exception.{ PROTOCOL_ERROR, INTERNAL_ERROR }
 
-import scala.annotation.tailrec
 import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.Duration
@@ -22,7 +21,7 @@ class BasicHttpStage(streamId: Int,
                            ec: ExecutionContext,
                       service: HttpService) extends TailStage[Http2Msg] {
 
-  import BasicHttpStage._
+  import Http2StageTools._
 
   import NodeMsg.{ DataFrame, HeadersFrame }
 
@@ -177,29 +176,3 @@ class BasicHttpStage(streamId: Int,
   }
 }
 
-private object BasicHttpStage {
-  // Request pseudo headers
-  val Method = ":method"
-  val Scheme = ":scheme"
-  val Path   = ":path"
-  val Authority = ":authority"
-
-  // Response pseudo header
-  val Status = ":status"
-  val Connection = "connection"
-  val TE = "te"
-  val ContentLength = "content-length"
-
-  def isLowerCase(str: String): Boolean = {
-    @tailrec
-    def go(i: Int): Boolean = {
-      if (i == str.length) true
-      else {
-        val ch = str.charAt(i)
-        if ('A' <= ch && ch <= 'Z') false
-        else go(i + 1)
-      }
-    }
-    go(0)
-  }
-}
